Refactored freely rotating / scaling UIElements into a separate Sprite class; removed rotation from the base class as it would not work correctly with element hierarchies in most cases.

+++ Engine/UI/ListView.cpp
@@ -728,7 +728,7 @@ void ListView::Expand(unsigned index, bool enable, bool recursive)
         bool visible = enable && expanded[indent - 1];
         item->SetVisible(visible);
 
-        if (indent >= expanded.Size())
+        if (indent >= (int)expanded.Size())
             expanded.Resize(indent + 1);
         expanded[indent] = visible && GetItemExpanded(item);
     }
